// src/helpers/log.type.enum.ts
export enum LogType {
  ERROR = 'ERROR',
  ACQUISITION = 'ACQUISITION',
  DEVOLUTION = 'DEVOLUTION',
  SOLD = 'SOLD',
  TRANSFER = 'TRANSFER',
  THEFT = 'THEFT',
  LOST = 'LOST',
  CONSUME = 'CONSUME',
  ADM_ADJUST = 'ADM_ADJUST',
  TRANSFER_BETWEEN_FILIALS = 'TRANSFER_BETWEEN_FILIALS',
  QUANTITY_CHECKED = 'QUANTITY_CHECKED',
  PRODUCT_CREATED = 'PRODUCT_CREATED',
  STORAGE_CREATED = 'STORAGE_CREATED',
}